requirements.txt
Django==2.2.28
pytz==2022.1
sqlparse==0.4.2
sorl-thumbnail==12.6.3
Pillow==9.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djoser==2.1.0
PyJWT==2.1.0
requests==2.26.0
